1.1.1 Fixed FPS tanking issue
The fps issue was caused by the "team menu" feature of the cheat due to looping over network objects way too many times when being activated.  This will need to be fixed in the future but is okay now due to it being easy to toggle and not a main feature.